+++ src/Plane.cpp
@@ -3,7 +3,7 @@
 #include "Ray.h"
 
 bool Plane::intersect(const Ray& ray, const double min_t, double& out_t,
-                      Eigen::Vector3d& out_n) const {
+                      Eigen::Vector3f& out_n) const {
     const double t = (this->point - ray.origin).dot(this->normal) /
                      ray.direction.dot(this->normal);
 
@@ -12,7 +12,7 @@ bool Plane::intersect(const Ray& ray, const double min_t, double& out_t,
                  std::remove_reference_t<decltype(out_t)>>::infinity())
         return false;
 
-    Eigen::Vector3d n = this->normal;
+    Eigen::Vector3f n = this->normal;
     if (n.dot(ray.direction) > 0) n = -n;
 
     out_t = t;
