+++ src/TriangleSoup.cpp
@@ -4,7 +4,7 @@
 #include "first_hit.h"
 
 bool TriangleSoup::intersect(const Ray& ray, const double min_t, double& t,
-                             Eigen::Vector3d& n) const {
+                             Eigen::Vector3f& n) const {
     int hit_id;
     return first_hit(ray, min_t, this->triangles, hit_id, t, n);
 }
